AppRepository: refactor out unrelated content
Create new classes that take responsibility for code that doesn't belong to
the AppRepository class and delete unused code. Also add stop() method to
SpeakController::AudioObserver interface to explicitly stop an ongoing
utterance.
